+++ web/src/hooks/api/useProjectDashboard.ts
@@ -1,0 +1,69 @@
+import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query";
+import { projectDashboardApi } from "@/services/api/project-dashboard.api";
+import type { CreatePurchaseOrderDTO, CreatePartyDTO, UpdatePurchaseOrderDTO } from "@/modules/operations/project-dashboard/helpers/projectDashboard.types";
+
+export const projectsDashboardKeys = {
+    all: ["projects-dashboard"] as const,
+    lists: () => [...projectsDashboardKeys.all, "list"] as const,
+    poParties: () => [...projectsDashboardKeys.all, "po-parties"] as const,
+    detail: (id: number) => [...projectsDashboardKeys.all, "detail", id] as const,
+    purchaseOrder: (id: number) => [...projectsDashboardKeys.all, "purchase-order", id] as const,
+};
+
+export const useProjectDashboardDetails = (id: number) => {
+    return useQuery({
+        queryKey: projectsDashboardKeys.detail(id),
+        queryFn: () => projectDashboardApi.getDashboardDetails(id),
+        enabled: !!id,
+    });
+};
+
+export const usePoParties = () => {
+    return useQuery({
+        queryKey: projectsDashboardKeys.poParties(),
+        queryFn: () => projectDashboardApi.getPoParties(),
+        staleTime: 5 * 60 * 1000,
+    });
+};
+
+export const usePurchaseOrderDetails = (id: number) => {
+    return useQuery({
+        queryKey: projectsDashboardKeys.purchaseOrder(id),
+        queryFn: () => projectDashboardApi.getPurchaseOrder(id),
+        enabled: !!id,
+    });
+};
+
+export const useCreatePurchaseOrder = () => {
+    const queryClient = useQueryClient();
+
+    return useMutation({
+        mutationFn: (data: CreatePurchaseOrderDTO) => projectDashboardApi.createPurchaseOrder(data),
+        onSuccess: () => {
+            queryClient.invalidateQueries({ queryKey: projectsDashboardKeys.all });
+        },
+    });
+};
+
+export const useUpdatePurchaseOrder = () => {
+    const queryClient = useQueryClient();
+
+    return useMutation({
+        mutationFn: ({ id, data }: { id: number; data: UpdatePurchaseOrderDTO }) =>
+            projectDashboardApi.updatePurchaseOrder(id, data),
+        onSuccess: () => {
+            queryClient.invalidateQueries({ queryKey: projectsDashboardKeys.all });
+        },
+    });
+};
+
+export const useCreatePoParty = () => {
+    const queryClient = useQueryClient();
+
+    return useMutation({
+        mutationFn: (data: CreatePartyDTO) => projectDashboardApi.createParty(data),
+        onSuccess: () => {
+            queryClient.invalidateQueries({ queryKey: projectsDashboardKeys.poParties() });
+        },
+    });
+};
